Move encoding logic into separate file
